+++ CHANGELOG.md
@@ -15,6 +15,9 @@ To compensate for the loss of Control key in Vim Emulation, we provide custom co
 -   Now you can toggle fullscreen mode by pressing F11 or clicking View->Full Screen. (#642, #660 and #670)
 -   In test cases, you can copy the output to the expected. (#688)
 -   Copy test cases between different tabs. (#688)
+-   Now settings are saved to disk as soon as new settings are applied. (#590 and #722)
+-   x86 release for 32-bit Windows. (#719 and #723)
+-   Open documentation links in the preferences window. (#733)
 
 ### Fixed